requirements.txt
selenium==4.21.0
pytest==8.2.2
python-dotenv==1.0.1
requests==2.32.3
